requirements.txt
streamlit
pandas
numpy
faiss-cpu
google-generativeai
python-dotenv
pydantic
